+++ src/lib/HAPHostNode.ts
@@ -13,21 +13,29 @@ import {
 } from 'hap-nodejs'
 import HapCategories from './types/HapCategories'
 import { SemVer } from 'semver'
+import { logger } from './logger'
 
 module.exports = (RED: NodeAPI, hostType: HostType) => {
-    const debug = require('debug')('NRCHKB:HAPHostNode')
-
     const MdnsUtils = require('./utils/MdnsUtils')()
 
     const init = function (this: HAPHostNodeType, config: HAPHostConfigType) {
         const self = this
+        const [logDebug, logError] = logger(
+            'HAPHostNode',
+            config.bridgeName,
+            self
+        )
+
         self.hostType = hostType
         RED.nodes.createNode(self, config)
 
         self.config = config
-
         self.name = config.bridgeName
-        debug('Setting name to ' + config.bridgeName)
+
+        if (!hostNameValidator(config.bridgeName)) {
+            logError('Host name is incorrect', false)
+            throw Error('Host name is incorrect')
+        }
 
         if (semver.parse(config.firmwareRev) == null) {
             config.firmwareRev = new SemVer('0.0.0')
@@ -76,15 +84,7 @@ module.exports = (RED: NodeAPI, hostType: HostType) => {
         const hostTypeName =
             self.hostType == HostType.BRIDGE ? 'Bridge' : 'Standalone Accessory'
 
-        debug(
-            'Creating ' +
-                hostTypeName +
-                " with name '" +
-                self.name +
-                "' and UUID '" +
-                hostUUID +
-                "'"
-        )
+        logDebug(`Creating ${hostTypeName} with UUID ${hostUUID}`)
 
         if (self.hostType == HostType.BRIDGE) {
             self.host = new Bridge(self.name, hostUUID)
@@ -93,30 +93,22 @@ module.exports = (RED: NodeAPI, hostType: HostType) => {
         }
 
         self.publish = function () {
-            debug(
-                'Publishing ' +
-                    hostTypeName +
-                    " with name '" +
-                    self.name +
-                    "', pin code '" +
-                    self.config.pinCode +
-                    (self.hostType == HostType.BRIDGE
-                        ? "' and " +
-                          self.host.bridgedAccessories.length +
-                          ' accessories.'
-                        : '.')
-            )
+            if (self.hostType == HostType.BRIDGE) {
+                logDebug(
+                    `Publishing ${hostTypeName} with pin code ${self.config.pinCode} and ${self.host.bridgedAccessories.length} accessories`
+                )
+            } else {
+                logDebug(
+                    `Publishing ${hostTypeName} with pin code ${self.config.pinCode}`
+                )
+            }
 
             if (
                 (self.config.port && self.config.port == 1880) ||
                 (self.mdnsConfig?.port && self.mdnsConfig?.port == 1880)
             ) {
-                self.error(
-                    'Cannot publish ' +
-                        hostTypeName +
-                        " '" +
-                        self.name +
-                        "' on port 1880 as it is reserved for node-red."
+                logError(
+                    `Cannot publish on ${hostTypeName} port 1880 as it is reserved for node-red`
                 )
                 self.published = false
                 return false
@@ -153,19 +145,9 @@ module.exports = (RED: NodeAPI, hostType: HostType) => {
 
         self.host.on('identify', function (paired: any, callback: () => any) {
             if (paired) {
-                debug(
-                    'Identify called on paired ' +
-                        hostTypeName +
-                        ' ' +
-                        self.name
-                )
+                logDebug(`Identify called on paired ${hostTypeName}`)
             } else {
-                debug(
-                    'Identify called on unpaired ' +
-                        hostTypeName +
-                        ' ' +
-                        self.name
-                )
+                logDebug(`Identify called on unpaired ${hostTypeName}`)
             }
 
             callback()
@@ -210,14 +192,17 @@ module.exports = (RED: NodeAPI, hostType: HostType) => {
             if (match) {
                 return match.join(':').substr(0, 17).toUpperCase()
             } else {
-                throw new Error(
-                    'match failed in macify process for padded string ' +
-                        paddedStr
+                throw Error(
+                    `match failed in macify process for padded string ${paddedStr}`
                 )
             }
         } else {
-            throw new Error('nodeId cannot be empty in macify process')
+            throw Error('nodeId cannot be empty in macify process')
         }
+    }
+
+    const hostNameValidator = function (hostName: string) {
+        return hostName ? /^[^.]{1,64}$/.test(hostName) : false
     }
 
     return {