+++ src/lib/HAPServiceNode.ts
@@ -4,11 +4,9 @@ import HAPServiceNodeType from './types/HAPServiceNodeType'
 import HAPHostNodeType from './types/HAPHostNodeType'
 import HostType from './types/HostType'
 import { uuid } from 'hap-nodejs'
-import UnifiedLogger from './utils/UnifiedLogger'
+import { logger } from './logger'
 
 module.exports = (RED: NodeAPI) => {
-    const debug = require('debug')('NRCHKB:HAPServiceNode')
-
     /**
      * Config override when user created services in old NRCHKB version
      */
@@ -17,11 +15,13 @@ module.exports = (RED: NodeAPI) => {
     ) {
         const self = this
 
+        const [, , logTrace] = logger('HAPServiceNode', self.config.name, self)
+
+        console.log(self.config)
+
         if (self.config.isParent === undefined) {
-            debug(
-                'nrchkbConfigCompatibilityOverride => self.config.isParent=' +
-                    self.config.isParent +
-                    ' value changed to true'
+            logTrace(
+                `nrchkbConfigCompatibilityOverride => self.config.isParent=${self.config.isParent} value changed to true`
             )
             // Services created in pre linked services era where working in 1.2 but due to more typescript in 1.3+ it started to cause some errors
             self.config.isParent = true
@@ -29,14 +29,10 @@ module.exports = (RED: NodeAPI) => {
 
         if (self.config.hostType === undefined) {
             // When moving from 1.2 to 1.3 hostType is not defined on homekit-service
-            if (self.config.isParent) {
-                debug(
-                    'nrchkbConfigCompatibilityOverride => self.config.hostType=' +
-                        self.config.hostType +
-                        ' value changed to HostType.BRIDGE'
-                )
-                self.config.hostType = HostType.BRIDGE
-            }
+            logTrace(
+                `nrchkbConfigCompatibilityOverride => self.config.hostType=${self.config.hostType} value changed to HostType.BRIDGE`
+            )
+            self.config.hostType = HostType.BRIDGE
         }
     }
 
@@ -46,17 +42,21 @@ module.exports = (RED: NodeAPI) => {
     ) {
         const self = this
         self.config = config
+        self.name = self.config.name
+
+        const [logDebug] = logger('HAPServiceNode', self.config.name, self)
+
         self.RED = RED
         self.publishTimers = {}
 
         nrchkbConfigCompatibilityOverride.call(self)
-        RED.nodes.createNode(self, config)
+        RED.nodes.createNode(self, self.config)
 
         const ServiceUtils = require('./utils/ServiceUtils')(self)
 
         new Promise<HAPServiceConfigType>((resolve) => {
-            if (config.waitForSetupMsg) {
-                debug(
+            if (self.config.waitForSetupMsg) {
+                logDebug(
                     'Waiting for Setup message. It should be of format {"payload":{"nrchkb":{"setup":{}}}}'
                 )
 
@@ -69,10 +69,10 @@ module.exports = (RED: NodeAPI) => {
                 })
 
                 self.handleWaitForSetup = (msg: Record<string, unknown>) =>
-                    ServiceUtils.handleWaitForSetup(config, msg, resolve)
+                    ServiceUtils.handleWaitForSetup(self.config, msg, resolve)
                 self.on('input', self.handleWaitForSetup)
             } else {
-                resolve(config)
+                resolve(self.config)
             }
         }).then((newConfig) => {
             init.call(self, newConfig)
@@ -86,35 +86,31 @@ module.exports = (RED: NodeAPI) => {
         const self = this
         self.config = config
 
+        const [logDebug, logError] = logger(
+            'HAPServiceNode',
+            self.config.name,
+            self
+        )
+
         const ServiceUtils = require('./utils/ServiceUtils')(self)
 
         if (self.config.isParent) {
-            debug('Starting Parent Service ' + config.name)
+            logDebug('Starting Parent Service')
             configure.call(self)
             self.configured = true
         } else {
+            const serviceType =
+                config.serviceName === 'CameraControl' ? 'Camera' : 'Linked'
+
             ServiceUtils.waitForParent()
                 .then(() => {
-                    debug(
-                        'Starting ' +
-                            (config.serviceName === 'CameraControl'
-                                ? 'Camera'
-                                : 'Linked') +
-                            ' Service ' +
-                            config.name
-                    )
+                    logDebug(`Starting  ${serviceType} Service`)
                     configure.call(self)
                     self.configured = true
                 })
                 .catch((error: any) => {
-                    UnifiedLogger.error(
-                        self,
-                        'Error while starting ' +
-                            (config.serviceName === 'CameraControl'
-                                ? 'Camera'
-                                : 'Linked') +
-                            ' Service',
-                        error
+                    logError(
+                        `Error while starting ${serviceType} Service due to ${error}`
                     )
                 })
         }
@@ -122,6 +118,12 @@ module.exports = (RED: NodeAPI) => {
 
     const configure = function (this: HAPServiceNodeType) {
         const self = this
+
+        const [logDebug, logError] = logger(
+            'HAPServiceNode',
+            self.config.name,
+            self
+        )
 
         const Utils = require('./utils')(self)
         const AccessoryUtils = Utils.AccessoryUtils
@@ -140,7 +142,7 @@ module.exports = (RED: NodeAPI) => {
             self.hostNode = RED.nodes.getNode(hostId) as HAPHostNodeType
 
             if (!self.hostNode) {
-                UnifiedLogger.error(self, 'Host Node not found')
+                logError('Host Node not found', false)
                 throw Error('Host Node not found')
             }
 
@@ -153,12 +155,14 @@ module.exports = (RED: NodeAPI) => {
             ) as HAPServiceNodeType
 
             if (!parentNode) {
+                logError('Parent Node not assigned', false)
                 throw Error('Parent Node not assigned')
             }
 
             self.parentService = parentNode.service
 
             if (!self.parentService) {
+                logError('Parent Service not assigned', false)
                 throw Error('Parent Service not assigned')
             }
 
@@ -176,22 +180,22 @@ module.exports = (RED: NodeAPI) => {
 
         // Look for existing Accessory or create a new one
         if (self.config.hostType == HostType.BRIDGE) {
-            // According to the HomeKit Accessory Protocol Specification the value
-            // of the fields Name, Manufacturer, Serial Number and Model must not
-            // change throughout the lifetime of an accessory. Because of that the
-            // accessory UUID will be generated based on that data to ensure that
-            // a new accessory will be created if any of those configuration values
-            // changes.
-            const accessoryUUID = uuid.generate(
-                'A' +
-                    self.id +
-                    self.name +
-                    self.config.manufacturer +
-                    self.config.serialNo +
-                    self.config.model
-            )
-
             if (self.config.isParent) {
+                // According to the HomeKit Accessory Protocol Specification the value
+                // of the fields Name, Manufacturer, Serial Number and Model must not
+                // change throughout the lifetime of an accessory. Because of that the
+                // accessory UUID will be generated based on that data to ensure that
+                // a new accessory will be created if any of those configuration values
+                // changes.
+                const accessoryUUID = uuid.generate(
+                    'A' +
+                        self.id +
+                        self.name +
+                        self.config.manufacturer +
+                        self.config.serialNo +
+                        self.config.model
+                )
+
                 self.accessory = AccessoryUtils.getOrCreate(
                     self.hostNode.host,
                     {
@@ -213,7 +217,7 @@ module.exports = (RED: NodeAPI) => {
             }
         } else {
             // We are using Standalone Accessory mode so no need to create new Accessory as we have "host" already
-            debug('Binding Service accessory as Standalone Accessory')
+            logDebug('Binding Service accessory as Standalone Accessory')
             self.accessory = self.hostNode.host
         }