+++ src/nodes/nrchkb.ts
@@ -3,26 +3,21 @@ import * as path from 'path'
 import semver from 'semver'
 import { HAPStorage } from 'hap-nodejs'
 import storage from 'node-persist'
+import { logger } from '../lib/logger'
 
-const debug = require('debug')('NRCHKB')
+const [logDebug, logError, logTrace] = logger()
 
 module.exports = (RED: NodeAPI) => {
     const requiredNodeVersion = '10.22.1'
     const nodeVersion = process.version
 
     if (semver.gte(nodeVersion, requiredNodeVersion)) {
-        debug(
-            'Node.js version requirement met. Required ' +
-                requiredNodeVersion +
-                '. Installed ' +
-                nodeVersion
+        logDebug(
+            `Node.js version requirement met. Required ${requiredNodeVersion}. Installed ${nodeVersion}`
         )
     } else {
-        throw new RangeError(
-            'Node.js version requirement not met. Required ' +
-                requiredNodeVersion +
-                '. Installed ' +
-                nodeVersion
+        throw RangeError(
+            `Node.js version requirement not met. Required ${requiredNodeVersion}. Installed ${nodeVersion}`
         )
     }
 
@@ -30,27 +25,33 @@ module.exports = (RED: NodeAPI) => {
 
     // Initialize our storage system
     if (RED.settings.available() && RED.settings.userDir) {
-        debug('RED settings available')
+        logDebug('RED settings available')
 
         const nrchkbStoragePath = path.resolve(RED.settings.userDir, 'nrchkb')
         storage.init({ dir: nrchkbStoragePath }).then(() => {
             // Initialize API
             API.init()
         })
-        debug('nrchkbStorage path set to ', nrchkbStoragePath)
+        logDebug(`nrchkbStorage path set to ${nrchkbStoragePath}`)
 
         const hapStoragePath = path.resolve(
             RED.settings.userDir,
             'homekit-persist'
         )
-        HAPStorage.setCustomStoragePath(hapStoragePath)
 
-        debug('HAPStorage path set to ', hapStoragePath)
+        try {
+            HAPStorage.setCustomStoragePath(hapStoragePath)
+            logDebug(`HAPStorage path set to ${hapStoragePath}`)
+        } catch (error) {
+            logDebug('HAPStorage already initialized')
+            logError('node-red restart highly recommended')
+            logTrace(error)
+        }
     } else {
-        debug('RED settings not available')
+        logDebug('RED settings not available')
     }
 
-    debug('Registering nrchkb type')
+    logDebug('Registering nrchkb type')
     RED.nodes.registerType('nrchkb', function (this: any, config) {
         RED.nodes.createNode(this, config)
     })